Adds "clicked" prop that run when row is clicked
When the row of a table is clicked, we would like to perform a redirect.

Allow the user to pass in a function that will be executed when the row
is clicked. It's up to the user to add the redirection within the method
as the method is passed the current item and its index.