feat(modal): Add Bootstrap V4 anticipated verticaly centered modal (#1246)
* feat(modal): Add Bootstrap V4 anticipated verticaly centered modal

Vertically centered modals will be coming in Bootstrap V4.beta.3

This adds a new prop `centered` which renders the modal centered vertically.

https://github.com/twbs/bootstrap/pull/24510

Note vertically centered modals are **not** suited for tall content modals, as there is a potential for the header (or top part of the modal) to become inaccessible).  They are better suited to one or two line content (i.e. confirmation dialogs).

The custom CSS in modal.vue can be removed once V4.beta.3 (or later) arrives

* update modal docs

* SpellCheck 

:scroll: